Added projection functions for 2-dim histograms
Added two functions project_on_x() and project_on_y() for both regular 2-dim Bin and
2-dim SparselyBin histograms.
They return a 1-dim Bin or SparselyBin histogram respectively, where the 2-dim histogram
has been projected on the x- or y-axis, i.e. has been integrated over the y- or x-axis.
So, one can fill a 2-dim histogram, and then extract the corresponding 1-dim histograms
from this.

Two more small things:

Also added a small bug-fix to the xy_ranges_grid() function of the class
SparselyTwoDimensionallyHistogramMethods, which now raises a KeyError if the histograms
is not filled.

In the plotmatplotlib() functions of TwoDimensionallyHistogramMethods and
SparselyTwoDimensionallyHistogramMethods, setting the x- and y-axis limits is
now done consistently after plotting the histograms.