Added plotting methods for Categorize type histograms
To make plotting work for Categorize histograms, I had to add a CategorizeHistogram
class to the file specialize.py, to add addImplicitMethods() to Categorize histograms,
in order to ensure that the plotting functions are recognized in python.
Similarly, in primitives/categorize.py I had to properly set the contentType of a Categorize object.
(I copied the example of SparselyHistogram where required.)

I only added matplotlib and ROOT plotting functions for Categorize histograms,
as I'm not very familiar with bokey.
The bokey plotting class for Categorize histograms is currently a dummy class.